NOJIRA-Billing-ledger-refactoring
- bin-billing-manager: Remove allowanceHandler dependency from billinghandler, accounthandler, listenhandler
- bin-billing-manager: Rewrite BillingEnd to use atomic BillingConsumeAndRecord
- bin-billing-manager: Simplify IsValidBalance to read BalanceToken/BalanceCredit directly from account
- bin-billing-manager: Change balance methods from float32 to int64
- bin-billing-manager: Remove allowance endpoints from listen handler
- bin-billing-manager: Update billing Create to set TransactionType and int64 rates

+++ bin-billing-manager/pkg/listenhandler/models/request/v1_accounts.go
@@ -12,12 +12,12 @@ type V1DataAccountsIDPUT struct {
 
 // V1DataAccountsIDBalanceAddForcePOST is rquest param define for POST /accounts/<account-id>/balance_add
 type V1DataAccountsIDBalanceAddForcePOST struct {
-	Balance float32 `json:"balance"`
+	Balance int64 `json:"balance"`
 }
 
 // V1DataAccountsIDBalanceSubtractForcePOST is rquest param define for POST /accounts/<account-id>/balance_subtract
 type V1DataAccountsIDBalanceSubtractForcePOST struct {
-	Balance float32 `json:"balance"`
+	Balance int64 `json:"balance"`
 }
 
 // V1DataAccountsIDIsValidBalancePOST is rquest param define for POST /accounts/<account-id>/is_valid_balance
